Start expending event manager + add performance HUD on OVR 0.6.0.1

+++ src/BtOgre.cpp
@@ -1028,7 +1028,7 @@ namespace BtOgre {
 	//------------------------------------------------------------------------------------------------
 	void DynamicLines::fillHardwareBuffers()
 	{
-	  int size = mPoints.size();
+	  size_t size = mPoints.size();
 
 	  prepareHardwareBuffers(size,0);
